Update description.md
I noticed this problem by solving the task on exercism.
Taking Rule # 3 literally, the "queen" test wrongly expects "eenquay" because in this case "qu" is not preceded by any consonant, so I think the correct result for this word should be "ueenqay". "according to the second rule. Perhaps maybe it would be better to change the expected "queen" test result.